requirements.txt
astroid==1.4.5
autopep8==1.2.2
beautifulsoup4==4.4.1
colorama==0.3.7
csscompressor==0.9.4
dj-database-url==0.4.1
Django==1.9.4
django-appconf==1.0.1
django-bower==5.1.0
django-compressor==2.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-filter==0.13.0
django-webpack-loader==0.3.0
djangorestframework==3.3.3
djangorestframework-camel-case==0.2.0
gunicorn==19.4.5
html5lib==0.9999999
lazy-object-proxy==1.2.1
Markdown==2.6.6
pep8==1.7.0
ply==3.8
psycopg2==2.6.1
pylint==1.5.5
pylint-django==0.7.1
pylint-plugin-utils==0.2.3
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
slimit==0.8.1
sqlparse==0.1.19
whitenoise==3.0
wrapt==1.10.7
